# artifacts/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "ApprovalStatusResponseV1": {
        "additionalProperties": false,
        "description": "Response model for GET /v1/approvals/{approval_id}",
        "properties": {
          "approval_id": {
            "description": "Approval request ID",
            "title": "Approval Id",
            "type": "string"
          },
          "correlation_id": {
            "description": "Correlation ID from original request",
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "description": "When the approval was created",
            "title": "Created At",
            "type": "string"
          },
          "requested_action_class": {
            "description": "Action class requiring approval",
            "title": "Requested Action Class",
            "type": "string"
          },
          "status": {
            "description": "Approval status (PENDING/APPROVED/DENIED)",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "status",
          "created_at",
          "requested_action_class",
          "correlation_id"
        ],
        "title": "ApprovalStatusResponseV1",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Autonomous Defense Mesh Organism v1 - Thin Vertical Slice API",
    "title": "ExoArmur ADMO v1 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/identity_containment/execute/{approval_id}": {
      "post": {
        "description": "Execute containment with approval",
        "operationId": "execute_approval_api_v2_identity_containment_execute__approval_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Approval"
      }
    },
    "/api/v2/identity_containment/intents/from_recommendation": {
      "post": {
        "description": "Create intent from recommendation",
        "operationId": "create_intent_from_recommendation_api_v2_identity_containment_intents_from_recommendation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Request",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Intent From Recommendation"
      }
    },
    "/api/v2/identity_containment/intents/{intent_id}": {
      "get": {
        "description": "Get intent details",
        "operationId": "get_intent_api_v2_identity_containment_intents__intent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intent"
      }
    },
    "/api/v2/identity_containment/recommendations": {
      "post": {
        "description": "Create containment recommendation",
        "operationId": "create_recommendation_api_v2_identity_containment_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Request",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Recommendation"
      }
    },
    "/api/v2/identity_containment/status": {
      "get": {
        "description": "Get containment status for a subject",
        "operationId": "get_containment_status_api_v2_identity_containment_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Containment Status"
      }
    },
...
    "/v1/approvals/{approval_id}": {
      "get": {
        "description": "Get approval status",
        "operationId": "get_approval_status_v1_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalStatusResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval Status"
      }
    },
    "/v1/approvals/{approval_id}/approve": {
      "post": {
        "description": "Approve an approval request",
        "operationId": "approve_approval_v1_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Approval"
      }
    },
    "/v1/approvals/{approval_id}/deny": {
      "post": {
        "description": "Deny an approval request",
        "operationId": "deny_approval_v1_approvals__approval_id__deny_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deny Approval"
      }
    },
    "/v1/audit/{correlation_id}": {
      "get": {
        "description": "Get audit records for a correlation ID",
        "operationId": "get_audit_records_v1_audit__correlation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "correlation_id",
            "required": true,
            "schema": {
              "title": "Correlation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Records"
      }
    },
    "/v1/telemetry/ingest": {
      "post": {
        "description": "Ingest telemetry event and process through ADMO loop",
        "operationId": "ingest_telemetry_v1_telemetry_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEventV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryIngestResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Telemetry"
      }
    }
  }
}